add default root property, cleanup plugin
