+++ packages/gasket-plugin-https/lib/create.js
@@ -1,0 +1,9 @@
+const { name, version } = require('../package.json');
+
+/** @type {import('@gasket/core').HookHandler<'create'>} */
+module.exports = async function create(gasket, { pkg, gasketConfig }) {
+  gasketConfig.addPlugin('pluginHttps', name);
+  pkg.add('dependencies', {
+    [name]: `^${version}`
+  });
+};
