+++ packages/gasket-plugin-https/lib/utils.js
@@ -1,3 +1,5 @@
+const proxy = require('http-proxy');
+
 /**
  * Provide port defaults
  * @param {string} env env property from gasket config
@@ -22,7 +24,43 @@ function portInUseError(errors) {
   );
 }
 
+/**
+ * Create a https proxy server for local development
+ * @param {import('@gasket/core').DevProxyConfig} opts Gasket instance
+ * @param {import('@gasket/plugin-logger').Logger} logger Gasket logger
+ */
+function startProxy(opts, logger) {
+  const { protocol = 'http', hostname = 'localhost', port = 8080, ...proxyOpts } = opts;
+  proxy.createServer(
+    proxyOpts
+  ).on('error', (e) => {
+    logger.error('Request failed to proxy:', e);
+  }).listen(
+    port
+  );
+
+  logger.info(`Proxy server started: ${protocol}://${hostname}:${port}`);
+}
+
+/**
+ * Get server options from the gasket config
+ * @param {import('@gasket/core').Gasket} gasket Gasket instance
+ * @returns {import('@gasket/core').ServerOptions} config
+ */
+function getRawServerConfig(gasket) {
+  const { hostname, http2, https, http, root } = gasket.config;
+  const rawConfig = {};
+  rawConfig.hostname = hostname;
+  rawConfig.root = root;
+  if (http) rawConfig.http = http;
+  if (https) rawConfig.https = https;
+  if (http2) rawConfig.http2 = http2;
+  return rawConfig;
+}
+
 module.exports = {
   getPortFallback,
-  portInUseError
+  portInUseError,
+  startProxy,
+  getRawServerConfig
 };
